backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.8.0
google-generativeai==0.4.1
python-dotenv==1.0.0
pydantic==2.5.0
tiktoken==0.5.2
pytest==7.4.3
httpx==0.25.2